apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.57.0
httpx==0.27.2
asyncpg==0.29.0
pgvector==0.3.2
python-multipart==0.0.12
pydantic==2.9.2
pydantic[email]==2.9.2
supabase==2.10.0
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
email-validator==2.1.0
